fix: job recovery bug.   1. SQL query was using ::INTERVAL which pgx couldn't encode from Go types   2. Go code was passing int64 but make_interval(secs => ) needs float64
  Changes:
  - Updated RecoverStaleJobs query to use make_interval(secs => )
  - Changed int64(...) to just w.config.StaleJobThreshold.Seconds() (returns float64

+++ internal/worker/worker.go
@@ -94,7 +94,7 @@ func (w *Worker) Stop() {
 // recoverStaleJobs finds jobs that have been running too long and resets them to pending.
 // This handles the case where a worker crashed while processing a job.
 func (w *Worker) recoverStaleJobs(ctx context.Context) error {
-	thresholdSeconds := int64(w.config.StaleJobThreshold.Seconds())
+	thresholdSeconds := w.config.StaleJobThreshold.Seconds()
 	count, err := w.queries.RecoverStaleJobs(ctx, thresholdSeconds)
 	if err != nil {
 		return fmt.Errorf("recover stale jobs: %w", err)
